+++ src/ignitus-Authentication/ignitus-ProfessorLogin/Components/login.js
@@ -4,8 +4,8 @@
 
 import React, { useState } from 'react';
 import { withErrorBoundary } from '../../../ignitus-Shared/Components/errorBoundary';
-import { statePayload } from '../../../ignitus-Shared/Components/Login/Constants';
-import SharedLogin from '../../../ignitus-Shared/Components/Login/Login';
+import { statePayload } from '../../../ignitus-Shared/Components/AuthenticationComponents/Login/Constants';
+import SharedLogin from '../../../ignitus-Shared/Components/AuthenticationComponents/Login/Login';
 
 const Login = () => {
   // Dummy Implementation.
